app/requirements.txt
# Server dependencies
fastapi
uvicorn[standard] # [standard] includes performance extras
Pillow
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
requests
